+++ src/utils/APIKey.js
@@ -1,0 +1,4 @@
+let APIKey = 'f18d3686ca97af45';
+
+export default APIKey;
+
